+++ rocketgram/api/chat.py
@@ -6,10 +6,11 @@
 from dataclasses import dataclass
 from typing import Dict, Optional
 
+from . import message
+from .chat_location import ChatLocation
 from .chat_permissions import ChatPermissions
 from .chat_photo import ChatPhoto
 from .chat_type import ChatType
-from . import message
 
 
 @dataclass(frozen=True)
@@ -30,6 +31,7 @@ class Chat:
     first_name: Optional[str]
     last_name: Optional[str]
     photo: Optional[ChatPhoto]
+    bio: Optional[str]
     description: Optional[str]
     invite_link: Optional[str]
     pinned_message: Optional['message.Message']
@@ -37,6 +39,8 @@ class Chat:
     slow_mode_delay: Optional[int]
     sticker_set_name: Optional[str]
     can_set_sticker_set: Optional[bool]
+    linked_chat_id: Optional[int]
+    location: Optional[ChatLocation]
 
     @classmethod
     def parse(cls, data: Optional[Dict]) -> Optional['Chat']:
@@ -44,7 +48,8 @@ class Chat:
             return None
 
         return cls(data['id'], ChatType(data['type']), data.get('title'), data.get('username'), data.get('first_name'),
-                   data.get('last_name'), ChatPhoto.parse(data.get('photo')), data.get('description'),
+                   data.get('last_name'), ChatPhoto.parse(data.get('photo')), data.get('bio'), data.get('description'),
                    data.get('invite_link'), message.Message.parse(data.get('pinned_message')),
                    ChatPermissions.parse(data.get('permissions')), data.get('slow_mode_delay'),
-                   data.get('sticker_set_name'), data.get('can_set_sticker_set'))
+                   data.get('sticker_set_name'), data.get('can_set_sticker_set'), data.get('linked_chat_id'),
+                   ChatLocation.parse(data.get('location')))